Jennifer's R dictionary
This has a few tutorials to help with some complex tasks in R including plotting, for loops, regex etc.  If you would like the data so that you can play around with these tutorials, please email me.  There are many ways to perform the same tasks in R these are just my common practices and methods.  I apologize in advance for the messiness of the document.  When I improve in markdown I will be sure to update this document and make it easy to read.  Until then, I suggest using the TOC to navigate.